+++ bindings/nostr-sdk-ffi/src/relay/mod.rs
@@ -23,10 +23,10 @@ pub use self::options::{ConnectionMode, RelayOptions, SubscribeOptions};
 pub use self::stats::RelayConnectionStats;
 pub use self::status::RelayStatus;
 use crate::database::events::Events;
+use crate::database::NostrDatabase;
 use crate::error::Result;
 use crate::negentropy::NegentropyItem;
 use crate::protocol::{ClientMessage, Event, EventId, Filter, RelayInformationDocument};
-use crate::NostrDatabase;
 
 #[derive(Record)]
 pub struct ReconciliationSendFailureItem {