+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 * ffi: drop support for `i686-linux-android` target ([Yuki Kishimoto])
 * ffi: remove `MockRelay` ([Yuki Kishimoto])
 * ffi: remove `RawEvent` and `RawEventRecord` ([Yuki Kishimoto])
+* ffi: remove `RelayPool` ([Yuki Kishimoto])
 
 ### Deprecated